requirements.txt
pyside6
pillow
numpy
